requirements.txt
asgiref==3.7.2
Django==4.2.3
django-cors-headers==4.2.0
django-cron==0.6.0
django-filter==23.2
django-redis==5.3.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
et-xmlfile==1.1.0
mysqlclient==2.2.0
numpy==1.25.2
openpyxl==3.1.2
Pillow==10.0.0
pyarrow==12.0.1
PyJWT==2.8.0
pylance==0.6.0
python-memcached==1.59
pytz==2023.3
redis==4.6.0
six==1.16.0
sqlparse==0.4.4
tzdata==2023.3
